+++ CHANGELOG.md
@@ -63,6 +63,7 @@
 
 - bugfix method lookup for nested named imported types
 - fix type signature of module types
+- only allow ``Int``, ``Signal``, ``String``, ``Regex`` type constant expression in case patterns
 
 ## [0.39.3] - 2025-10-31